banco_dados/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.0.0
setuptools==74.1.2
